Basic C Major Chords
Adding KeyboardState visualisation using the standard midi keyboard stuff from JUCE. Tweaked the styles a little bit to see what was easily possible. Also wrote some simple logic to convert notes to Major Triads.

NOTE: This works ok as long as multiple chords are not simultaneously being pressed. If they are simultaneous, there may be some issues when the different chords share notes and the handling of the note off messages. This is something that needs to be looked at in the future.